+++ NEWS.md
@@ -1,5 +1,11 @@
 <!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
 
+# filenamr development
+
+* Refactor to consistently pass `call` parameter when using `cli_abort()` for error checking.
+* Update `read_exif()` to drop `geometry` parameter and drop sf from Suggests for package.
+* Expand tests for `get_data_dir()`.
+
 # filenamr 0.1.0.9001 (2023-03-20)
 
 * Export `check_path_fileext()` function.